feat(overtime): server-side balance guard for payouts (#426)
Extract the overtime/monthly-stats engine out of ReportController into a
reusable OvertimeCalculationService (getMonthlyStats/getRangeStats moved 1:1,
new getNetOvertimeMinutes). OvertimePayoutService::create() now validates a
payout against the net overtime balance and rejects amounts that would push it
negative, closing the direct-POST bypass of the frontend cap.

Depends on OvertimePayoutMapper (not the service) to avoid a dependency cycle.
Behavior of the existing report endpoints is unchanged: the migrated overtime
tests run against the extracted service and stay green.

Fixes #426
