+++ lib/Controller/ReportController.php
@@ -17,6 +17,7 @@ use OCA\WorkTime\Db\TimeEntryMapper;
 use OCA\WorkTime\Service\AbsenceService;
 use OCA\WorkTime\Service\EmployeeService;
 use OCA\WorkTime\Service\HolidayService;
+use OCA\WorkTime\Service\OvertimeCalculationService;
 use OCA\WorkTime\Service\OvertimePayoutService;
 use OCA\WorkTime\Service\PdfService;
 use OCA\WorkTime\Service\PermissionService;
@@ -50,6 +51,7 @@ class ReportController extends BaseController {
         private WorkScheduleService $workScheduleService,
         private YearlyCarryoverService $carryoverService,
         private OvertimePayoutService $payoutService,
+        private OvertimeCalculationService $overtimeCalc,
         private ProjectService $projectService,
         private IL10N $l,
     ) {
@@ -85,7 +87,7 @@ class ReportController extends BaseController {
             $holidays = $this->holidayService->findByMonth($year, $month, $employee->getFederalState());
 
             // Calculate statistics
-            $stats = $this->calculateMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
+            $stats = $this->overtimeCalc->getMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
 
             // Per-day labor-law warnings (#338): minimum break (§4 ArbZG) and max
             // daily hours, evaluated across all entries of each day.
@@ -549,7 +551,7 @@ class ReportController extends BaseController {
             $absences = $this->absenceService->findByEmployeeAndMonth($employeeId, $year, $month);
             $holidays = $this->holidayService->findByMonth($year, $month, $employee->getFederalState());
 
-            $stats = $this->calculateMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
+            $stats = $this->overtimeCalc->getMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
 
             $pdfContent = $this->pdfService->generateMonthlyReport(
                 $employee,
@@ -610,7 +612,7 @@ class ReportController extends BaseController {
             $absences = $this->absenceService->findByEmployeeAndDateRange($employeeId, $start, $end);
             $holidays = $this->holidayService->findHolidaysInRange($start, $end, $employee->getFederalState());
 
-            $stats = $this->calculateRangeStats($employee, $start, $end, $timeEntries, $absences, $holidays);
+            $stats = $this->overtimeCalc->getRangeStats($employee, $start, $end, $timeEntries, $absences, $holidays);
 
             $pdfContent = $this->pdfService->generateRangeReport(
                 $employee,
@@ -662,7 +664,7 @@ class ReportController extends BaseController {
             $absences = $allAbsences[$empId] ?? [];
             $holidays = $this->getHolidaysCached($year, $month, $employee->getFederalState());
 
-            $stats = $this->calculateMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
+            $stats = $this->overtimeCalc->getMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
             $statusSummary = $allStatusSummaries[$empId] ?? ['draft' => 0, 'submitted' => 0, 'approved' => 0, 'rejected' => 0];
 
             $report[] = [
@@ -762,7 +764,7 @@ class ReportController extends BaseController {
                 $absences = $absencesByMonth[$month];
                 $holidays = $this->getHolidaysCached($year, $month, $employee->getFederalState());
 
-                $stats = $this->calculateMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
+                $stats = $this->overtimeCalc->getMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
                 $statusSummary = $allStatusByMonth[$month][$empId] ?? ['draft' => 0, 'submitted' => 0, 'approved' => 0, 'rejected' => 0];
 
                 // Determine dominant status
@@ -859,7 +861,7 @@ class ReportController extends BaseController {
                 $absences = $this->absenceService->findByEmployeeAndMonth($employeeId, $year, $month);
                 $holidays = $this->holidayService->findByMonth($year, $month, $employee->getFederalState());
 
-                $stats = $this->calculateMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
+                $stats = $this->overtimeCalc->getMonthlyStats($employee, $year, $month, $timeEntries, $absences, $holidays);
 
                 $monthlyData[] = [
                     'month' => $month,
@@ -967,470 +969,4 @@ class ReportController extends BaseController {
         return $this->workScheduleService->countWorkingDays($absence->getEmployeeId(), $start, $end, []);
     }
 
-    /**
-     * Returns the current date (today, time zeroed). Wrapped in a method so tests
-     * can pin "today" to a fixed date and verify the proportional Soll/overtime logic.
-     */
-    protected function currentDate(): DateTime {
-        return new DateTime('today');
-    }
-
-    /**
-     * Whether the given day has activity that makes it count toward the proportional
-     * Soll: a time entry on that day or an approved absence covering it.
-     *
-     * @param object[] $timeEntries
-     * @param object[] $absences
-     */
-    private function hasActivityOnDay(DateTime $day, array $timeEntries, array $absences): bool {
-        $dayStr = $day->format('Y-m-d');
-        foreach ($timeEntries as $entry) {
-            $entryDate = $entry->getDate();
-            if ($entryDate instanceof DateTime && $entryDate->format('Y-m-d') === $dayStr) {
-                return true;
-            }
-        }
-        foreach ($absences as $absence) {
-            if ($absence->isApproved()
-                && $absence->getStartDate() <= $day
-                && $absence->getEndDate() >= $day) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Statistics for a custom [start, end] range (#102). Unlike the monthly
-     * calculation there is no "until today" proportional logic: a custom-period
-     * timesheet covers a fixed, user-chosen span, so Soll/Ist/Saldo are computed
-     * over the whole range.
-     *
-     * @param TimeEntry[] $timeEntries
-     * @param Absence[] $absences
-     * @param Holiday[] $holidays
-     * @return array<string, mixed>
-     */
-    private function calculateRangeStats(
-        Employee $employee,
-        DateTime $startDate,
-        DateTime $endDate,
-        array $timeEntries,
-        array $absences,
-        array $holidays
-    ): array {
-        $employeeId = $employee->getId();
-
-        // Clip the range to the employment period.
-        $entryDate = $employee->getEntryDate();
-        $exitDate = $employee->getExitDate();
-        if ($entryDate !== null && $startDate < $entryDate) {
-            $startDate = clone $entryDate;
-        }
-        if ($exitDate !== null && $endDate > $exitDate) {
-            $endDate = clone $exitDate;
-        }
-
-        $entryCount = count($timeEntries);
-        if ($startDate > $endDate) {
-            return [
-                'workingDays' => 0, 'holidayCount' => count($holidays),
-                'paidAbsenceDays' => 0, 'targetReductionDays' => 0, 'compensatoryDays' => 0,
-                'absenceDays' => 0, 'adjustedTargetMinutes' => 0, 'actualMinutes' => 0,
-                'overtimeMinutes' => 0, 'entryCount' => $entryCount,
-            ];
-        }
-
-        $workingDays = $this->workScheduleService->countWorkingDays($employeeId, $startDate, $endDate, $holidays);
-        $targetMinutes = $this->workScheduleService->calculateTargetMinutes($employeeId, $startDate, $endDate, $holidays);
-
-        $targetReductionTypes = [\OCA\WorkTime\Db\Absence::TYPE_UNPAID];
-        $overtimeConsumingTypes = [\OCA\WorkTime\Db\Absence::TYPE_COMPENSATORY];
-
-        $paidAbsenceMinutes = 0;
-        $paidAbsenceDays = 0;
-        $targetReductionMinutes = 0;
-        $targetReductionDays = 0;
-        $compensatoryDays = 0;
-
-        foreach ($absences as $absence) {
-            if (!$absence->isApproved()) {
-                continue;
-            }
-            $aStart = $absence->getStartDate() < $startDate ? $startDate : $absence->getStartDate();
-            $aEnd = $absence->getEndDate() > $endDate ? $endDate : $absence->getEndDate();
-            if ($aStart > $aEnd) {
-                continue;
-            }
-            $scope = $absence->getScopeValue();
-            $days = $this->workScheduleService->countWorkingDays($employeeId, $aStart, $aEnd, $holidays) * $scope;
-            $minutes = $this->calculateAbsenceMinutes($employeeId, $aStart, $aEnd, $scope, $holidays);
-
-            if (in_array($absence->getType(), $targetReductionTypes, true)) {
-                $targetReductionDays += $days;
-                $targetReductionMinutes += $minutes;
-            } elseif (in_array($absence->getType(), $overtimeConsumingTypes, true)) {
-                $compensatoryDays += $days;
-            } else {
-                $paidAbsenceDays += $days;
-                $paidAbsenceMinutes += $minutes;
-            }
-        }
-
-        $adjustedTargetMinutes = $targetMinutes - $targetReductionMinutes;
-
-        $workedMinutes = 0;
-        foreach ($timeEntries as $entry) {
-            $workedMinutes += $entry->getWorkMinutes();
-        }
-        $actualMinutes = $workedMinutes + $paidAbsenceMinutes;
-        $overtimeMinutes = $actualMinutes - $adjustedTargetMinutes;
-
-        return [
-            'workingDays' => $workingDays,
-            'holidayCount' => count($holidays),
-            'paidAbsenceDays' => $paidAbsenceDays,
-            'targetReductionDays' => $targetReductionDays,
-            'compensatoryDays' => $compensatoryDays,
-            'absenceDays' => $paidAbsenceDays + $targetReductionDays + $compensatoryDays,
-            'adjustedTargetMinutes' => $adjustedTargetMinutes,
-            'actualMinutes' => $actualMinutes,
-            'overtimeMinutes' => $overtimeMinutes,
-            'entryCount' => $entryCount,
-        ];
-    }
-
-    private function calculateMonthlyStats(
-        Employee $employee,
-        int $year,
-        int $month,
-        array $timeEntries,
-        array $absences,
-        array $holidays
-    ): array {
-        $startDate = new DateTime("$year-$month-01");
-        $monthEndDate = (clone $startDate)->modify('last day of this month');
-        $today = $this->currentDate();
-        $employeeId = $employee->getId();
-
-        // Clip to employee's employment period
-        $entryDate = $employee->getEntryDate();
-        $exitDate = $employee->getExitDate();
-
-        // Month entirely before employment start → return zeros
-        if ($entryDate !== null && $monthEndDate < $entryDate) {
-            return $this->zeroMonthStats(workingDaysOverride: 0);
-        }
-
-        // Month entirely after employment end → return zeros
-        if ($exitDate !== null && $startDate > $exitDate) {
-            return $this->zeroMonthStats(workingDaysOverride: 0);
-        }
-
-        // Clip start/end to employment period
-        if ($entryDate !== null && $startDate < $entryDate) {
-            $startDate = clone $entryDate;
-        }
-        if ($exitDate !== null && $monthEndDate > $exitDate) {
-            $monthEndDate = clone $exitDate;
-        }
-
-        // Determine if this is a future month (hasn't started yet)
-        $isFutureMonth = $startDate > $today;
-
-        // For future months: return zeros for Soll/Ist/Überstunden
-        if ($isFutureMonth) {
-            $workingDaysMonth = $this->workScheduleService->countWorkingDays($employeeId, $startDate, $monthEndDate, $holidays);
-            $monthlyTargetMinutes = $this->workScheduleService->calculateTargetMinutes($employeeId, $startDate, $monthEndDate, $holidays);
-
-            // Count planned absences for display only
-            $absenceDaysMonth = 0;
-            foreach ($absences as $absence) {
-                if ($absence->isApproved()) {
-                    $absenceStart = $absence->getStartDate() < $startDate ? $startDate : $absence->getStartDate();
-                    $absenceEnd = $absence->getEndDate() > $monthEndDate ? $monthEndDate : $absence->getEndDate();
-                    $absenceDaysMonth += $this->workScheduleService->countWorkingDays($employeeId, $absenceStart, $absenceEnd, $holidays);
-                }
-            }
-
-            // dailyMinutes approximation for display
-            $dailyMinutes = $workingDaysMonth > 0 ? (int)round($monthlyTargetMinutes / $workingDaysMonth) : 0;
-
-            return [
-                'workingDays' => $workingDaysMonth,
-                'workingDaysMonth' => $workingDaysMonth,
-                'workingDaysUntilToday' => 0,
-                'holidayCount' => count($holidays),
-                'paidAbsenceDays' => $absenceDaysMonth,
-                'targetReductionDays' => 0,
-                'compensatoryDays' => 0,
-                'absenceDays' => $absenceDaysMonth,
-                'dailyMinutes' => $dailyMinutes,
-                'targetMinutes' => 0,
-                'monthlyTargetMinutes' => 0,
-                'adjustedTargetMinutes' => 0,
-                'adjustedMonthlyTargetMinutes' => 0,
-                'workedMinutes' => 0,
-                'workedHours' => 0,
-                'paidAbsenceMinutes' => 0,
-                'paidAbsenceHours' => 0,
-                'actualMinutes' => 0,
-                'actualHours' => 0,
-                'overtimeMinutes' => 0,
-                'overtimeHours' => 0,
-                'entryCount' => 0,
-                'isFutureMonth' => true,
-            ];
-        }
-
-        // Determine if this is the current month
-        $isCurrentMonth = $year === (int)$today->format('Y') && $month === (int)$today->format('n');
-
-        // For "up to today" calculations.
-        // The running day only counts toward the proportional Soll once it has activity
-        // (a time entry today or an approved absence covering today). Without that,
-        // today is excluded so the balance shows no spurious morning deficit.
-        $endDateForActual = $monthEndDate;
-        if ($isCurrentMonth && $today < $monthEndDate) {
-            $endDateForActual = $this->hasActivityOnDay($today, $timeEntries, $absences)
-                ? $today
-                : (clone $today)->modify('-1 day');
-        }
-
-        // If today is excluded and it is the first (working) day of the month, the
-        // "until today" range falls before the month start -> no proportional Soll yet.
-        $hasProportionalRange = $endDateForActual >= $startDate;
-
-        // Count working days using schedule-aware service
-        $workingDaysMonth = $this->workScheduleService->countWorkingDays($employeeId, $startDate, $monthEndDate, $holidays);
-        $workingDaysUntilToday = $hasProportionalRange
-            ? $this->workScheduleService->countWorkingDays($employeeId, $startDate, $endDateForActual, $holidays)
-            : 0.0;
-
-        // Calculate target minutes using schedule-aware service
-        $monthlyTargetMinutes = $this->workScheduleService->calculateTargetMinutes($employeeId, $startDate, $monthEndDate, $holidays);
-        $proportionalTargetMinutes = $hasProportionalRange
-            ? $this->workScheduleService->calculateTargetMinutes($employeeId, $startDate, $endDateForActual, $holidays)
-            : 0;
-
-        // dailyMinutes approximation for display
-        $dailyMinutes = $workingDaysMonth > 0 ? (int)round($monthlyTargetMinutes / $workingDaysMonth) : 0;
-
-        // Process absences: paid (credit Ist) vs. target-reducing (Soll) vs. compensatory.
-        $paidAbsenceMinutesMonth = 0;
-        $paidAbsenceDaysMonth = 0;
-        $targetReductionMinutesMonth = 0;
-        $targetReductionDaysMonth = 0;
-        $compensatoryDaysMonth = 0;
-
-        $paidAbsenceMinutesUntilToday = 0;
-        $paidAbsenceDaysUntilToday = 0;
-        $targetReductionMinutesUntilToday = 0;
-        $targetReductionDaysUntilToday = 0;
-        $compensatoryDaysUntilToday = 0;
-
-        // Types that reduce the target (Soll) instead of crediting work time (Ist).
-        $targetReductionTypes = [
-            \OCA\WorkTime\Db\Absence::TYPE_UNPAID,
-        ];
-
-        // Compensatory time (Freizeitausgleich) is neither credited to the Ist nor
-        // deducted from the Soll: the day stays a target day with no work credited, so
-        // the overtime balance drops by one daily target when FZA is taken (#149, #186).
-        $overtimeConsumingTypes = [
-            \OCA\WorkTime\Db\Absence::TYPE_COMPENSATORY,
-        ];
-
-        foreach ($absences as $absence) {
-            if ($absence->isApproved()) {
-                $absenceStart = $absence->getStartDate();
-                $absenceEnd = $absence->getEndDate();
-                $absenceScope = $absence->getScopeValue();
-
-                // For full month display
-                if ($absenceStart <= $monthEndDate) {
-                    $monthAbsenceStart = $absenceStart < $startDate ? $startDate : $absenceStart;
-                    $monthAbsenceEnd = $absenceEnd > $monthEndDate ? $monthEndDate : $absenceEnd;
-                    $daysInMonth = $this->workScheduleService->countWorkingDays($employeeId, $monthAbsenceStart, $monthAbsenceEnd, $holidays);
-                    $effectiveDays = $daysInMonth * $absenceScope;
-
-                    // Calculate absence minutes per day using actual schedule
-                    $absenceMinutes = $this->calculateAbsenceMinutes($employeeId, $monthAbsenceStart, $monthAbsenceEnd, $absenceScope, $holidays);
-
-                    if (in_array($absence->getType(), $targetReductionTypes, true)) {
-                        $targetReductionDaysMonth += $effectiveDays;
-                        $targetReductionMinutesMonth += $absenceMinutes;
-                    } elseif (in_array($absence->getType(), $overtimeConsumingTypes, true)) {
-                        // FZA: counted as an absence day, but not credited to the Ist.
-                        $compensatoryDaysMonth += $effectiveDays;
-                    } else {
-                        $paidAbsenceDaysMonth += $effectiveDays;
-                        $paidAbsenceMinutesMonth += $absenceMinutes;
-                    }
-                }
-
-                // For actual/overtime calculation: only count absences up to today
-                if ($hasProportionalRange && $absenceStart <= $endDateForActual) {
-                    $actualAbsenceStart = $absenceStart < $startDate ? $startDate : $absenceStart;
-                    $actualAbsenceEnd = $absenceEnd > $endDateForActual ? $endDateForActual : $absenceEnd;
-                    $daysUntilToday = $this->workScheduleService->countWorkingDays($employeeId, $actualAbsenceStart, $actualAbsenceEnd, $holidays);
-                    $effectiveDaysUntilToday = $daysUntilToday * $absenceScope;
-
-                    $absenceMinutesUntilToday = $this->calculateAbsenceMinutes($employeeId, $actualAbsenceStart, $actualAbsenceEnd, $absenceScope, $holidays);
-
-                    if (in_array($absence->getType(), $targetReductionTypes, true)) {
-                        $targetReductionDaysUntilToday += $effectiveDaysUntilToday;
-                        $targetReductionMinutesUntilToday += $absenceMinutesUntilToday;
-                    } elseif (in_array($absence->getType(), $overtimeConsumingTypes, true)) {
-                        // FZA: not credited to the Ist -> overtime decreases by the daily target.
-                        $compensatoryDaysUntilToday += $effectiveDaysUntilToday;
-                    } else {
-                        $paidAbsenceDaysUntilToday += $effectiveDaysUntilToday;
-                        $paidAbsenceMinutesUntilToday += $absenceMinutesUntilToday;
-                    }
-                }
-            }
-        }
-
-        // Adjust targets for unpaid leave (compensatory time deliberately keeps the target).
-        $adjustedMonthlyTargetMinutes = $monthlyTargetMinutes - $targetReductionMinutesMonth;
-        $adjustedProportionalTargetMinutes = $proportionalTargetMinutes - $targetReductionMinutesUntilToday;
-
-        // Sum actual work minutes from time entries
-        $workedMinutes = 0;
-        foreach ($timeEntries as $entry) {
-            $workedMinutes += $entry->getWorkMinutes();
-        }
-
-        // Effective actual = worked + paid absences (up to today)
-        $actualMinutes = $workedMinutes + $paidAbsenceMinutesUntilToday;
-
-        // Calculate overtime against proportional target (not full month)
-        $overtimeMinutes = $actualMinutes - $adjustedProportionalTargetMinutes;
-
-        return [
-            // Full month values (for display in Monatsübersicht)
-            'workingDays' => $workingDaysMonth,
-            'workingDaysMonth' => $workingDaysMonth,
-            'workingDaysUntilToday' => $workingDaysUntilToday,
-            'holidayCount' => count($holidays),
-            'paidAbsenceDays' => $paidAbsenceDaysMonth,
-            'targetReductionDays' => $targetReductionDaysMonth,  // Unpaid leave
-            'compensatoryDays' => $compensatoryDaysMonth,
-            'absenceDays' => $paidAbsenceDaysMonth + $targetReductionDaysMonth + $compensatoryDaysMonth,
-            'dailyMinutes' => $dailyMinutes,
-
-            // Target values
-            'targetMinutes' => $adjustedMonthlyTargetMinutes,
-            'monthlyTargetMinutes' => $monthlyTargetMinutes,
-            'adjustedTargetMinutes' => $adjustedProportionalTargetMinutes,
-            'adjustedMonthlyTargetMinutes' => $adjustedMonthlyTargetMinutes,
-
-            // Actual values (up to today)
-            'workedMinutes' => $workedMinutes,
-            'workedHours' => round($workedMinutes / 60, 2),
-            'paidAbsenceMinutes' => $paidAbsenceMinutesUntilToday,
-            'paidAbsenceHours' => round($paidAbsenceMinutesUntilToday / 60, 2),
-            'actualMinutes' => $actualMinutes,
-            'actualHours' => round($actualMinutes / 60, 2),
-
-            // Overtime (proportional)
-            'overtimeMinutes' => $overtimeMinutes,
-            'overtimeHours' => round($overtimeMinutes / 60, 2),
-            'entryCount' => count($timeEntries),
-            'isFutureMonth' => false,
-        ];
-    }
-
-    private function zeroMonthStats(int $workingDaysOverride = 0): array {
-        return [
-            'workingDays' => $workingDaysOverride,
-            'workingDaysMonth' => $workingDaysOverride,
-            'workingDaysUntilToday' => 0,
-            'holidayCount' => 0,
-            'paidAbsenceDays' => 0,
-            'targetReductionDays' => 0,
-            'compensatoryDays' => 0,
-            'absenceDays' => 0,
-            'dailyMinutes' => 0,
-            'targetMinutes' => 0,
-            'monthlyTargetMinutes' => 0,
-            'adjustedTargetMinutes' => 0,
-            'adjustedMonthlyTargetMinutes' => 0,
-            'workedMinutes' => 0,
-            'workedHours' => 0,
-            'paidAbsenceMinutes' => 0,
-            'paidAbsenceHours' => 0,
-            'actualMinutes' => 0,
-            'actualHours' => 0,
-            'overtimeMinutes' => 0,
-            'overtimeHours' => 0,
-            'entryCount' => 0,
-            'isFutureMonth' => false,
-        ];
-    }
-
-    /**
-     * Calculate absence minutes using actual schedule for each day.
-     */
-    private function calculateAbsenceMinutes(int $employeeId, DateTime $start, DateTime $end, float $scope, array $holidays): int {
-        $holidayMap = [];
-        foreach ($holidays as $holiday) {
-            $holidayMap[$holiday->getDate()->format('Y-m-d')] = $holiday;
-        }
-
-        $totalMinutes = 0;
-        $current = clone $start;
-        while ($current <= $end) {
-            $dateStr = $current->format('Y-m-d');
-            $dayMinutes = $this->workScheduleService->getDailyMinutesForDate($employeeId, $current);
-
-            if ($dayMinutes > 0 && !isset($holidayMap[$dateStr])) {
-                $totalMinutes += (int)round($dayMinutes * $scope);
-            }
-            $current->modify('+1 day');
-        }
-
-        return $totalMinutes;
-    }
-
-    /**
-     * Count working days (Mon-Fri) excluding holidays
-     * Holiday scope determines how much of the day is free:
-     * - scope = 1.0: full holiday = 0 working days
-     * - scope = 0.5: half holiday = 0.5 working days remaining
-     */
-    private function countWorkingDays(DateTime $startDate, DateTime $endDate, array $holidays): float {
-        // Build holiday lookup: date => Holiday object
-        $holidayMap = [];
-        foreach ($holidays as $holiday) {
-            $holidayMap[$holiday->getDate()->format('Y-m-d')] = $holiday;
-        }
-
-        $workingDays = 0.0;
-        $current = clone $startDate;
-
-        while ($current <= $endDate) {
-            $dayOfWeek = (int)$current->format('N');
-            $dateStr = $current->format('Y-m-d');
-
-            // Only count Monday-Friday
-            if ($dayOfWeek < 6) {
-                if (isset($holidayMap[$dateStr])) {
-                    // Holiday exists - add remaining working portion
-                    // scope = 1.0 (full holiday) → 0 working days
-                    // scope = 0.5 (half holiday) → 0.5 working days
-                    $holiday = $holidayMap[$dateStr];
-                    $workingDays += (1.0 - $holiday->getScopeValue());
-                } else {
-                    // Regular working day
-                    $workingDays += 1.0;
-                }
-            }
-
-            $current->modify('+1 day');
-        }
-
-        return $workingDays;
-    }
 }
