+++ tests/Unit/Controller/ProportionalOvertimeTodayTest.php
@@ -5,25 +5,17 @@ declare(strict_types=1);
 namespace OCA\WorkTime\Tests\Unit\Controller;
 
 use DateTime;
-use OCA\WorkTime\Controller\ReportController;
 use OCA\WorkTime\Db\Absence;
-use OCA\WorkTime\Db\AbsenceMapper;
 use OCA\WorkTime\Db\Employee;
-use OCA\WorkTime\Db\TimeEntryMapper;
+use OCA\WorkTime\Db\OvertimePayoutMapper;
 use OCA\WorkTime\Service\AbsenceService;
 use OCA\WorkTime\Service\EmployeeService;
 use OCA\WorkTime\Service\HolidayService;
-use OCA\WorkTime\Service\PdfService;
-use OCA\WorkTime\Service\PermissionService;
-use OCA\WorkTime\Service\ProjectService;
+use OCA\WorkTime\Service\OvertimeCalculationService;
 use OCA\WorkTime\Service\TimeEntryService;
 use OCA\WorkTime\Service\WorkScheduleService;
 use OCA\WorkTime\Service\YearlyCarryoverService;
-use OCA\WorkTime\Service\OvertimePayoutService;
-use OCP\IL10N;
-use OCP\IRequest;
 use PHPUnit\Framework\TestCase;
-use ReflectionMethod;
 
 /**
  * Regression test for #251.5: the running ("current") day must not produce a
@@ -40,31 +32,23 @@ class ProportionalOvertimeTodayTest extends TestCase {
 
 	private const DAILY = 480;
 
-	private function makeController(WorkScheduleService $schedule, string $today): ReportController {
-		$controller = new class(
-			$this->createMock(IRequest::class),
-			'tester',
-			$this->createMock(TimeEntryService::class),
-			$this->createMock(TimeEntryMapper::class),
-			$this->createMock(AbsenceMapper::class),
-			$this->createMock(AbsenceService::class),
-			$this->createMock(EmployeeService::class),
-			$this->createMock(HolidayService::class),
-			$this->createMock(PermissionService::class),
-			$this->createMock(PdfService::class),
+	private function makeService(WorkScheduleService $schedule, string $today): OvertimeCalculationService {
+		$service = new class(
 			$schedule,
 			$this->createMock(YearlyCarryoverService::class),
-			$this->createMock(OvertimePayoutService::class),
-			$this->createMock(ProjectService::class),
-			$this->createMock(IL10N::class),
-		) extends ReportController {
+			$this->createMock(OvertimePayoutMapper::class),
+			$this->createMock(EmployeeService::class),
+			$this->createMock(TimeEntryService::class),
+			$this->createMock(AbsenceService::class),
+			$this->createMock(HolidayService::class),
+		) extends OvertimeCalculationService {
 			public string $pinnedToday = '';
 			protected function currentDate(): DateTime {
 				return new DateTime($this->pinnedToday);
 			}
 		};
-		$controller->pinnedToday = $today;
-		return $controller;
+		$service->pinnedToday = $today;
+		return $service;
 	}
 
 	/**
@@ -72,11 +56,9 @@ class ProportionalOvertimeTodayTest extends TestCase {
 	 * @param object[] $absences
 	 */
 	private function overtimeFor(string $today, array $entries, array $absences): int {
-		$controller = $this->makeController($this->schedule(), $today);
-		$method = new ReflectionMethod(ReportController::class, 'calculateMonthlyStats');
-		$method->setAccessible(true);
+		$service = $this->makeService($this->schedule(), $today);
 		// June 2026 is the "current" month relative to the pinned today.
-		$stats = $method->invoke($controller, $this->employee(), 2026, 6, $entries, $absences, []);
+		$stats = $service->getMonthlyStats($this->employee(), 2026, 6, $entries, $absences, []);
 		return $stats['overtimeMinutes'];
 	}
