+++ tests/Unit/Controller/ReportControllerTest.php
@@ -21,6 +21,7 @@ use OCA\WorkTime\Service\TimeEntryService;
 use OCA\WorkTime\Service\WorkScheduleService;
 use OCA\WorkTime\Service\YearlyCarryoverService;
 use OCA\WorkTime\Service\OvertimePayoutService;
+use OCA\WorkTime\Service\OvertimeCalculationService;
 use OCP\IL10N;
 use OCP\IRequest;
 use PHPUnit\Framework\TestCase;
@@ -58,6 +59,7 @@ class ReportControllerTest extends TestCase {
             $this->createMock(WorkScheduleService::class),
             $this->createMock(YearlyCarryoverService::class),
             $this->createMock(OvertimePayoutService::class),
+            $this->createMock(OvertimeCalculationService::class),
             $this->projectService,
             $this->createMock(IL10N::class),
         );
@@ -146,7 +148,8 @@ class ReportControllerTest extends TestCase {
             $this->createMock(AbsenceMapper::class), $this->createMock(AbsenceService::class),
             $this->employeeService, $this->createMock(HolidayService::class), $perm,
             $this->createMock(PdfService::class), $this->createMock(WorkScheduleService::class),
-            $this->createMock(YearlyCarryoverService::class), $this->createMock(OvertimePayoutService::class), $this->projectService,
+            $this->createMock(YearlyCarryoverService::class), $this->createMock(OvertimePayoutService::class),
+            $this->createMock(OvertimeCalculationService::class), $this->projectService,
             $this->createMock(IL10N::class),
         );
